requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
waitress==3.0.2
Werkzeug==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.23
greenlet>=3.0.3
psycopg2-binary
python-dotenv